Win Condition Works. Sorta...
The text is a bit weird and I can't figure out how to reset the start
scene.
When it resets now it still has the input from the last scene so you can
just keep hitting 'A' to play again.

Addded score reset method to player class
Added a reset all method to game manager class